+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod common;
 pub mod utils;
+pub mod notes;
