+++ tests/naming_test.rs
@@ -1,12 +1,13 @@
 use core::slice;
-use miden_client::{note::{NoteExecutionMode, NoteTag}, testing::NoteBuilder, transaction::AccountInterface};
+use miden_client::{note::{NoteExecutionMode, NoteTag}, testing::NoteBuilder, transaction::{AccountInterface, OutputNote}};
 use miden_crypto::Felt;
-use miden_testing::{MockChainBuilder};
+use miden_testing::{Auth, MockChain, MockChainBuilder, TransactionContextBuilder};
 use rand_chacha::ChaCha20Rng;
 use rand::{Rng, SeedableRng};
 use miden_objects::note::{NoteType};
-use crate::utils::{create_account, create_naming_account, create_naming_library, get_note_code};
-mod utils;
+use midenid_contracts::utils::{create_account, create_naming_account, create_naming_library};
+use midenid_contracts::notes::{get_note_code, create_naming_initialize_note};
+
 
 
 
@@ -14,6 +15,33 @@ mod utils;
 // https://github.com/0xMiden/miden-base/blob/719ff03d1482e6ce2ad4e986f59ec7b9a8ddf962/crates/miden-testing/src/kernel_tests/tx/test_fpi.rs#L515
 
 #[tokio::test]
+async fn test_naming_init() -> anyhow::Result<()> {
+    let mut builder = MockChain::builder();
+
+    let owner_account = builder.add_existing_wallet(Auth::BasicAuth)?;
+    let treasury_account= builder.add_existing_wallet(Auth::BasicAuth)?;
+
+    let naming_account = create_naming_account();
+
+    let initialize_input_note = create_naming_initialize_note(owner_account, treasury_account, naming_account.clone()).await.unwrap();
+
+    builder.add_note(OutputNote::Full(initialize_input_note.clone()));
+
+    let mock_chain = builder.build()?;
+
+    let tx_inputs = mock_chain.get_transaction_inputs(naming_account.clone(), None, &[initialize_input_note.id()], &[])?;
+
+    let tx_context = TransactionContextBuilder::new(naming_account.clone())
+        .account_seed(None)
+        .tx_inputs(tx_inputs)
+        .build()?;
+
+    let _exec_tx = tx_context.execute().await?;
+    Ok(())
+}
+
+#[tokio::test]
+#[ignore = "reason"]
 async fn test_naming_initialize() -> anyhow::Result<()>{
     let owner_account = create_account()?;
     let treasury_account = create_account()?;
